+++ src/sys/os/linux/structs/mod.rs
@@ -5,8 +5,13 @@
 
 mod sigaction;
 mod stat;
-mod termios;
 mod timespec;
-
 #[allow(unused)]
-pub use {sigaction::*, stat::*, termios::*, timespec::*};
+pub use {sigaction::*, stat::*, timespec::*};
+
+#[cfg(feature = "term")]
+crate::items! {
+    #[cfg_attr(nightly_doc, doc(cfg(feature = "term")))]
+    mod termios;
+    pub use termios::*;
+}